#85 Pipes for displaying date and time of messages created
