+++ Cooper/ClientApp/src.v2/app/components/chats-list/chats-list.component.ts
@@ -10,18 +10,6 @@ const fieldPermissibleMaxLength: number = 30;
 })
 export class ChatsListComponent implements OnInit {
 
-  public monthNames: string[] = [
-    'Jan', 'Feb', 'Mar',
-    'Apr', 'May', 'Jun', 'Jul',
-    'Aug', 'Sep', 'Oct',
-    'Nov', 'Dec'
-  ];
-
-  public dayNames: string[] = [
-    'Mon', 'Tue', 'Wed',
-    'Thu', 'Fri', 'Sat', 'Sun'
-  ];
-
   @Input() public currentSessionUserId: number;
   @Input() public newMessageBlockOpened: boolean;
   @Input() public chatsList: Chat[];
@@ -65,51 +53,5 @@ export class ChatsListComponent implements OnInit {
     } else {
       return messageContent;
     }
-  }
-
-  public getLastMessageSentTime(chat: Chat): string {
-
-    const date: Date = new Date(Date.parse(chat.messages[chat.messages.length - 1].createDate.toString()));
-    const dateNow: Date = new Date();
-    let messageTime: string;
-
-    if (dateNow.getFullYear() !== date.getFullYear()) {
-      const year: string = (date.getFullYear() % 100).toString();
-      const month: string = date.getMonth().toString();
-      const dayOfMonth: string = date.getDate().toString();
-
-      messageTime = `${dayOfMonth}.${month}.${year}`;
-
-    } else if (dateNow.getMonth() !== date.getMonth() ||
-                dateNow.getDate() - date.getDate() > 6 ||
-                dateNow.getDay() < date.getDay()) {
-
-      const month: string = this.monthNames[date.getMonth()];
-      const dayOfMonth: string = date.getDate().toString();
-
-      messageTime = `${month} ${dayOfMonth}`;
-
-    } else if (dateNow.getDay() !== date.getDay()) {
-
-      const dayOfWeek: string = this.dayNames[date.getDay() - 1];
-
-      messageTime = `${dayOfWeek}`;
-
-    } else {
-      let hours: string = date.getHours().toString();
-      let minutes: string = date.getMinutes().toString();
-
-      if (hours.length === 1) {
-        hours = '0' + hours;
-      }
-
-      if (minutes.length === 1) {
-        minutes = '0' + minutes;
-      }
-
-      messageTime = `${hours}:${minutes}`;
-    }
-
-    return messageTime;
   }
 }
