+++ Cooper/ClientApp/src.v2/app/pipes/chat-pipes/message-date-time.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { MessageDateTimePipe } from './message-date-time.pipe';
+
+describe('MessageDateTimePipe', () => {
+  it('create an instance', () => {
+    const pipe = new MessageDateTimePipe();
+    expect(pipe).toBeTruthy();
+  });
+});
