+++ Cooper/ClientApp/src.v2/app/pipes/chat-pipes/message-time.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { MessageTimePipe } from './message-time.pipe';
+
+describe('MessageTimePipe', () => {
+  it('create an instance', () => {
+    const pipe = new MessageTimePipe();
+    expect(pipe).toBeTruthy();
+  });
+});
